src/requirements.txt
pyzmq==22.0.3
click==8.0.1
#mininet==2.3.0